requirements.txt
pyyaml
pytz
schema
GitPython
